backend/requirements.txt
# Core Django and DRF
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Database
pymongo==4.6.3

# Google Drive API
google-api-python-client==2.120.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# Environment and Configuration
python-dotenv==1.0.0
django-cors-headers==4.3.1

# Development Tools
django-extensions==3.2.3
drf-yasg==1.21.7

# Testing and Development
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==4.1.0
coverage==7.4.4
factory-boy==3.3.0
faker==23.2.1

# Code Quality
black==24.2.0
isort==5.13.2
flake8==7.0.0
pre-commit==3.6.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
django-compressor==4.4

# Caching and Background Tasks
django-redis==5.4.0
celery==5.3.4

# Monitoring (Optional)
sentry-sdk==1.40.0

# Utilities
requests==2.31.0
Pillow==10.2.0

# ==============================================================================
# Multimedia RAG Engine
# ==============================================================================

# Video Processing
opencv-python==4.9.0.80
numpy==1.26.4

# Embeddings
sentence-transformers==2.5.1

# PyTorch (CPU build — use torch+cu118 for GPU)
torch==2.2.1
torchvision==0.17.1

# LLM / query pipeline
ollama==0.1.7

# Florence-2 frame captioning
transformers==4.40.0
# timm is required by the Florence-2 vision encoder
timm==0.9.16
# einops is required by Florence-2 attention layers
einops==0.8.2

# Person re-identification (install separately if needed: torchreid)
# torchreid==1.4.0
# gdown==5.1.0